added console UI presenter - a bit work left on other elements

+++ BattleOfTheShipsData/Exceptions/InvalidShotException.cs
@@ -8,7 +8,7 @@ namespace BattleOfTheShipsData.Exceptions
 {
 	public class InvalidShotException : MapPointException
 	{
-		public InvalidShotException(int x, int y) : base(x,y,$"Invalid Shot at ({x},x{y}).")
+		public InvalidShotException(long x, long y) : base(x,y,$"Invalid Shot at ({x},x{y}).")
 		{
 
 		}
